LPS-71164 Move methods to BaseSourceCheck
